Turn inno_gen into a struct + set head on init
Turns the closure pattern of getting innovation ids into a regular
struct of which we pass mutable references around to. This also allows
us to read the current head of the pathfunc so that we know where we
should start generating innovation ids from

Adds a param for the initial inno id head, so that we don't end up
generating duplicate innovation ids when creating a new pathfunc for
new generations of genomes.

For now, get rid of the default arc<mutex<_>> on the head and path -> id
map. arc<mutex<_>>-ing the entire struct may ( probably ) be useful in
the future, so a new_arc method is added to put an arc<mutex<_>> around
the entire stuct. Time will tell if this is the exact pattern that ends
up getting used