no-hash weight-diff and disjoint-excess counters
Since we are sorting connections by inno id when we generate them in
crossover, we know that by the time they get to avg-weight-diff and
disjoint-excess-count they'll be in order. Because of this.
We can therefore skip all hashing / set operations, and iter each pair
once.

We walk through the l and r connecitons at the same time. If either
yields nothing, we know that we're done getting weight diffs. We find
each next overlapping inno id and track their absolute difference.

While I'm at it, make the weight-diff tests print the diff if it's
not what we expect