An interface for event hooks while evolving
Add an interface for setting up functions ( or, hooks ) to fire when
EvolutionTargets are satisfied during evolution runs.

Hooks return an empty ControlFlow, which allows us to stop evolution
when a hook wants to, which in turn allows us to implement what was
accomplished by target in terms of a hook.