Decouple network from genome trait
Create trait FromGenome that can be implemented on a network, where some
network may take some genome and construct itself. In a sense, we are
returning the responsibility of creating the network back to the network
itself. This allows any Network construction from any Genome to be
implemented.

Remove network type and method from trait, moving them to their own
ToNetwork, allowing some Genome to construct any Network that implements
FromGenome for it. Automatically imlement ToNetwork in that case