Semi-standardize param mutation
Create a macro for implementing mutate_params on a connection, given an
events descriptor and corresponding probabilities. It's still a WIP as
it assumes the prameter generation range ( and, that params are all f64
). It uses a PARAM_PERTURB_FAC ( which replaces WEIGHT_MUTATION_FAC ),
moving that into the Connection trait.

The Param[Perturb, Replace] events are removed, picking a u64
probability branch over them - it makes sense that any param mutation
will be either a replacement or a perturbation, so we don't need that
overhead.