Hashless population allocation
Calculate specie allocations ( that is, the desired population of some
specie relative to the total population, based on their fitness )
without a hash table store and lookup. Because we know that the
species_fitted vec is going to be lined up ot the species vec that we
recieve, we can simply zip them together and iter out the actual
populaiton allocations for each specie once we know the total adjusted
fitness

This then allows us to apply a filter_map to exclude species whos
populations are zero ( I wonder if this step is even necessary? )
directly