FittedGroup for groups of fitted T
Create a trait FittedGroup, implementing some methods for any fitted
group - that is any group of (T, f64). In general, this allows us to
reason about any collection of genomes ( or, whatever ) as if it were a
specie without the need for an actual construction of said speice.

Implement it for Specie, since its members is a fitted group, pulling
fit_adjusted out of the regular impl.

Update reproduce, where the fittest member is found manualy, to simply
get .fittest()