Generate open paths that may include a point
Amend the Genome::open_path such that a path point to include, either
`From` or `To` a particular node, is a part of the open path returned.
If the inclusion specified cannot be satisfied, return None.

Add some helper methods on reuccurrent to get bound/unbound from/to
nodes, and use them to split up open_path a little

If we generate an invalid path, print it